requirements.txt
streamlit==1.33.0
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
chromadb==0.4.24
openai==1.14.3
pypdf==4.2.0
tiktoken==0.7.0
pydantic==1.10.13
httpx==0.27.0
python-dotenv==1.0.1